refactor: split main.py to separate files; improve overall project structure
