+++ .gitignore
@@ -3,3 +3,5 @@
 *.env
 __pycache__/
 .vscode/
+logs/
+databases/
