+++ ChangeLog.md
@@ -1,4 +1,4 @@
-# Чейнджлог "Лучшего друга сливок"
+# Ченджлог "Лучшего друга сливок"
 
 ## **v1.0**
 
@@ -12,7 +12,7 @@
 
 -----
 
-Разделение на парну/непарну недели
+Разделение на чётную/нечётную недели
 
 -----
 
@@ -59,5 +59,13 @@
 Расписание на субботу
 
 Улучшение стабильности работы
+
+-----
+
+## **v5.0**
+
+-----
+
+Добавлены задания
 
 -----
