--- assignments.json
@@ -1,19 +1,0 @@
-{
-    "assignments": [
-        {
-            "subject": "Some hueta predmet 2",
-            "task": "aaaa ya ne dolbaeb",
-            "deadline": "2024-09-20"
-        },
-        {
-            "subject": "Пенис",
-            "task": "Пенис",
-            "deadline": "2025-02-28"
-        },
-        {
-            "subject": "Я пидорас",
-            "task": "А может и не пидорас",
-            "deadline": "2024-10-15"
-        }
-    ]
-}
