+++ pyproject.toml
@@ -1,0 +1,22 @@
+[project]
+name = "telegram-ai221-sentinel"
+version = "0.1.0"
+description = ""
+authors = [
+    { name = "YoshkinKit", email = "nikita.zhmak8872@gmail.com" }, 
+    { name = "killmeridze", email = "kurtcobainuarethebest@gmail.com" },
+]
+readme = "README.md"
+requires-python = ">=3.12,<4.0"
+dependencies = [
+    "schedule (>=1.2.2,<2.0.0)",
+    "loguru (>=0.7.3,<0.8.0)",
+    "python-dotenv (>=1.0.1,<2.0.0)",
+    "requests (>=2.32.3,<3.0.0)",
+    "pytelegrambotapi (>=4.26.0,<5.0.0)"
+]
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
