+++ src/data/assignments.json
@@ -1,0 +1,3 @@
+{
+    "assignments": []
+}
