+++ src/data/localization.json
